requirements.txt
pandas
pygal_maps_world
cairosvg
imageio